[!!!][BUGFIX] Fix type detection of gender property
The getGender() method in the Actor entity now returns
a GenderEnum. Otherwise the Type detection gets confused
and thinks $gender is an integer.

Also the gender value 0 is now "unspecified" and 3 is "male".

For actors of type "company" the getGender() method always
return gender "unspecified".

This requires a database update!

UPDATE tx_theaterinfo_domain_model_actor
   SET gender = 3
 WHERE gender = 0
   AND type = 0;